Fix warning.

+++ Packet++/src/PostgresLayer.cpp
@@ -770,7 +770,7 @@ namespace pcpp
 			column.name.assign(iter, nameEnd);
 
 			// +1 offset because the nameEnd is currently at the null terminator of column name.
-			if (end - nameEnd < sizeof(::internal::PostgresColumnFixedData) + 1)
+			if (end - nameEnd < static_cast<std::ptrdiff_t>(sizeof(::internal::PostgresColumnFixedData) + 1))
 			{
 				columns.push_back(column);
 				break;
